requirements.txt
requests>=2.31.0
pyyaml>=6.0
python-telegram-bot==13.15
python-dotenv>=1.0.0
pytz>=2024.1
retrying>=1.3.4
vnstock==0.2.9.2.3
